Fixed NAV_SYS_SOB

+++ kernel/arch/ppc/timer.c
@@ -96,7 +96,7 @@ void pok_arch_decr_int (void)
 void ja_time_init (void)
 {
   time_inter = pok_bsp.timebase_freq / POK_TIMER_FREQUENCY;
-  printf("Timer interval: %lu\n", time_inter);
+  printf("Timer interval: %lu\n", (long unsigned int)time_inter);
   time_first = time_last = get_timebase ();
   set_decrementer();
